Module Tezos_oxymeter

Main module for the Tezos_oxymeter. It exposes the different modules and wrap the behaviour for the energy consumption observation.

module Report : sig ... end

Utility used by the entire library to represent a report. Once it's created, it can't be ereased.

module Observer : sig ... end

This module provides modules and functions to represent the interactions with the outside world.

module Metrics : sig ... end

This module defines a mechanism to gather information from observer during an observation process.

module Args : sig ... end

This module provides variables usable as arguments for the Arg module.