Module Tezos_oxymeter.Args

This module provides variables usable as arguments for the Arg module.

Gets arguments

val want_time : unit -> bool

want_time () returns true if the time metric is wanted.

val want_power : unit -> string option

want_power () returns Some (energy type) if the power metric is wanted. Else, it returns None.

val want_signal : unit -> bool

want_signal () returns true if a signal handler is requested.

val want_path : unit -> string

want_path () returns the path value. Default is /tmp/oxymeter-report.

Specifications

val power_spec : string * Stdlib.Arg.spec * string

Specification for a potential power argument in command line.

val time_spec : string * Stdlib.Arg.spec * string

Specification for a potential time argument in command line.

val signal_spec : string * Stdlib.Arg.spec * string

Specification for a potential signal argument in command line.

val path_spec : string * Stdlib.Arg.spec * string

Specification for a potential path argument in command line.