Module Tezos_oxymeter.Metrics

This module defines a mechanism to gather information from observer during an observation process.

Measure

Measures are defined as a unit you want to get when you observer the consumption.

module type MEASURE = sig ... end

Module that represents a measure unit.

module EnergyMeasure : MEASURE

Module with informations about the energy measure.

module TimeMeasure : MEASURE

Module with informations about the time measure .

Metrics

Metrics are the way to record measures until the end of the program or when a signal is triggered.

module type METRICS = sig ... end

Module that represents an instrument to take and save mesures.

module MakeMetrics : functor (M : MEASURE) -> METRICS

Functor to create an instrument to get METRICS from a MEASURE unit.

module TimeMetrics : METRICS

Module to gather metrics about time.

module EnergyMetrics : METRICS

Module to gather metrics about consumption.