Parameter MakeMetrics.1-M

type t

Abstract type to represent a measure.

val file : string

The name of file where the measure are exported.

val init : string list -> unit

Initialize the instrument.

val getMeasure : unit -> t Lwt.t

Gets a measure.

val diff : t -> t -> t Lwt.t

diff x y computes the difference between x and y.

val to_string_lwt : t -> string Lwt.t

Returns the representation of the t value as a string.

val to_ezjsonm_lwt : t -> Ezjsonm.value Lwt.t

Returns the representation of the t value as an Ezjsonm.value.